requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
jinja2
psycopg2-binary
python-dotenv
